Added feature to concatenate images to pdf
This script wasn't working in my case (had to concatenate 2 JPGS) so I added the functionality to concatenate images too, thanks to img2pdf. I tried to keep the script working as before, I only added temporary converted files that are removed after the joined file is created with pdftk.